automatically create nondigest non versioned files on asset precompile
The way non digest assets are created changed a little bit from always
copying the whole to only copying changed files on each run.  With this,
the task can automatically be invoked on each precompile run and only
copies something if needed.

The only Rails config change to make this gem work now is to add a
`Rails.application.config.assets.precompile += %w( webshims/* )` to an
initializer.